Reestruturação das funções estáticas dentro de 'Utils.py'
