+++ WebScraping/MyLibs/calendar_class.py
@@ -1,9 +1,6 @@
 import requests
 from bs4 import BeautifulSoup
 from datetime import datetime
-import pdfplumber
-import json
-import os
 import re
 from typing import List, Dict, Any
 
@@ -71,27 +68,6 @@ class Calendar:
         else:
             print('\033[31mNenhuma data válida encontrada nos textos\033[0m')
             return None, None
-
-    def download_pdf(self, url, filename):
-        response = requests.get(url)
-        if response.status_code == 200:
-            with open(filename, 'wb') as file:
-                file.write(response.content)
-            print(f'PDF baixado e salvo como {filename}')
-        else:
-            print(f'\033[31mFalha ao baixar o PDF. Status code: {response.status_code}\033[0m')
-
-    def extract_text_from_pdf(self, pdf_path):
-        text = ''
-        with pdfplumber.open(pdf_path) as pdf:
-            for page in pdf.pages:
-                text += page.extract_text() + '\n'
-        return text
-
-    def save_data_to_json(self, data, json_path):
-        with open(json_path, 'w', encoding='utf-8') as json_file:
-            json.dump(data, json_file, ensure_ascii=False, indent=4)
-        print(f'Dados salvos em {json_path}.')
 
     def upload_to_firestore(self, data):
         doc_ref = self.db.collection('APIs').document('Calendar')
@@ -178,11 +154,3 @@ class Calendar:
             "content": modified_events,
         }
         return formatted_data
-
-    def clean_up_files(self, *filenames, debug=True):
-        for filename in filenames:
-            if os.path.exists(filename):
-                os.remove(filename)
-                if debug:
-                    print('\033[36m-->\033[0m', end=' ')
-                    print(f'Arquivo {filename} removido com sucesso')
